Do major refactor regarding collisions between Me and FriendZones
FriendZones colliders are now MeshColliders, meaning 3D physics are now used.
Previously, each zone had a listener on its polygon collider that would notify the MeFriendZonesHandler class. In turn, this class would determine the zone the player is in (since the zone colliders weren't concave). Then the information was passed on to FriendZonesController which would update the friendzones colours accordingly.
Now, only a listener attached to the player expects the friendzones collisions, which it turn calls the script attached to the collider, which calls a method directly from the corresponding FriendZone class.